+++ apps/backend/src/modules/auth/auth.module.ts
@@ -1,11 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
+import { StorageModule } from '../../infrastructure/storage/storage.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 
 @Module({
   imports: [
+    StorageModule,
     JwtModule.registerAsync({
       global: true,
       imports: [ConfigModule],